src/website/requirements.txt
alembic==1.14.0
asgiref==3.8.1
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.4
django-author==1.2.0
django-cleanup==9.0.0
django-clearcache==1.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-js-asset==2.2.0
django-multiselectfield==0.1.13
django-nested-admin==4.1.1
django-quill-editor==0.1.42
django-rest-swagger==2.2.0
django-storages==1.14.4
django-timezone==0.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
environ==1.0
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
invoke==2.2.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
Mako==1.3.7
MarkupSafe==3.0.2
numpy==2.1.3
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-monkey-business==1.1.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
rsa==4.9
setuptools-git==1.2
simplejson==3.19.3
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
waitress==3.0.2
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.8.2
